Loginsystem & Models für Deck und Trucks

+++ project/Skateshop/Skateshop/Controllers/UsersController.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Skateshop.Data;
 using Skateshop.Models;
@@ -37,14 +36,15 @@ namespace Skateshop.Controllers
         {
             if (ModelState.IsValid && !_context.User.Any(u => u.Username.Equals(user.Username)))
             {
-                _context.Add(new User {
+                _context.Add(new User
+                {
                     Username = user.Username,
                     Password = Cipher.Encrypt(user.Password)
                 });
                 await _context.SaveChangesAsync();
 
                 _authService.Login(HttpContext, user);
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction("Index", "Home");
             }
             ViewBag.Error = $"The username \"{user.Username}\" already exists.";
             return View(user);
@@ -67,7 +67,7 @@ namespace Skateshop.Controllers
         {
             if (ModelState.IsValid && _authService.Login(HttpContext, user))
             {
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction("Index", "Home");
             }
             ViewBag.Error = "Wrong username or password";
             return View(user);
@@ -95,7 +95,11 @@ namespace Skateshop.Controllers
         // GET: Users
         public async Task<IActionResult> Index()
         {
-            return View(await _context.User.ToListAsync());
+            if (_authService.IsAdmin(HttpContext))
+            {
+                return View(await _context.User.ToListAsync());
+            }
+            return NotFound();
         }
 
         // GET: Users/Details/5
@@ -116,32 +120,10 @@ namespace Skateshop.Controllers
             return View(user);
         }
 
-        // GET: Users/Create
-        public IActionResult Create()
-        {
-            return View();
-        }
-
-        // POST: Users/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Username,Password")] User user)
-        {
-            if (ModelState.IsValid)
-            {
-                _context.Add(user);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            return View(user);
-        }
-
         // GET: Users/Edit/5
         public async Task<IActionResult> Edit(long? id)
         {
-            if (id == null)
+            if (id == null || !_authService.IsAdmin(HttpContext))
             {
                 return NotFound();
             }
@@ -161,7 +143,7 @@ namespace Skateshop.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(long id, [Bind("Id,Username,Password")] User user)
         {
-            if (id != user.Id)
+            if (id != user.Id || !_authService.IsAdmin(HttpContext))
             {
                 return NotFound();
             }
@@ -192,7 +174,7 @@ namespace Skateshop.Controllers
         // GET: Users/Delete/5
         public async Task<IActionResult> Delete(long? id)
         {
-            if (id == null)
+            if (id == null || !_authService.IsAdmin(HttpContext))
             {
                 return NotFound();
             }
@@ -212,6 +194,10 @@ namespace Skateshop.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(long id)
         {
+            if (!_authService.IsAdmin(HttpContext))
+            {
+                return NotFound();
+            }
             var user = await _context.User.FindAsync(id);
             _context.User.Remove(user);
             await _context.SaveChangesAsync();
